% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_domain.R
\name{geom_domain}
\alias{geom_domain}
\title{A 'ggplot2' geom to draw protein domains}
\usage{
geom_domain(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  height = grid::unit(5, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} opbject giving the height of the rectangle
representing the domain. Defaults to 5 mm.}
}
\description{
\code{geom_domain()} draws a geom representing a protein domain.
}
\details{
The domain is drawn as rounded rectangle, sitting on the molecular
backbone.The rectangle has a fill which is light grey by default.

The range of the domain on the molecular backbone is set with the \code{xmin} and
\code{xmax} aesthetics. The molecular backbone that it is associated with is set
with the \code{y} aesthetic.
}
\section{Aesthetics}{

\itemize{
\item xmin,xmax (required; start and end positions of the domain)
\item y (required; the molecular backbone)
\item alpha
\item colour
\item linetype
\item linewidth
\item fill
}
}

\examples{

ggplot2::ggplot(demolin_domains, aes(xmin = start, xmax = end, y = 1)) +
  geom_domain(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_domain_label]{geom_domain_label()}}
}
