% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_domain_label.R
\name{geom_domain_label}
\alias{geom_domain_label}
\title{A 'ggplot2' geom to add labels to domains}
\usage{
geom_domain_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = FALSE,
  place = "inside",
  align = "centre",
  height = grid::unit(4, "mm"),
  label_height = grid::unit(3, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{place}{Where to draw the label, either 'inside' the domain (the
default) or 'above' it}

\item{align}{How the text label should be aligned relative to the ends of
the domain Default is 'centre'; other options are 'left' and 'right'.}

\item{height}{\code{grid::unit()} object giving the height of the label above the
molecular backbone, if place is 'outside'. Defaults to 4 mm.}

\item{label_height}{\code{grid::unit()} object giving the height of the label
text. Defaults to 3 mm.}
}
\description{
\code{geom_domain_label()} adds labels to domains drawn with
\code{geom_domain()}.
}
\details{
Standard 'ggplot2' aesthetics for text are supported (see Aesthetics).
}
\section{Aesthetics}{

\itemize{
\item xmin,xmax (required; start and end positions of the domain)
\item y (required; the molecular backbone)
\item label (required; the label text)
\item colour
\item linewidth
\item alpha
\item family
\item fontface
\item angle
}
}

\examples{

ggplot2::ggplot(demolin_domains, aes(xmin = start, xmax = end, y = 1,
                                     label = name)) +
  geom_domain(inherit.aes = TRUE) +
  geom_domain_label(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_domain]{geom_domain()}}
}
