% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_helix.R
\name{geom_helix}
\alias{geom_helix}
\title{A 'ggplot2' geom to draw alpha helices}
\usage{
geom_helix(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  height = grid::unit(5, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} opbject giving the height of the helix.
Defaults to 5 mm.}
}
\description{
\code{geom_helix()} draws a geom representing an alpha helix in a protein
secondary structure diagram.
}
\details{
The helix is drawn as a horizontal helix, 'wrapped around' the molecular
backbone.

The range of the helix on the molecular backbone is set with the \code{xmin} and
\code{xmax} aesthetics. The molecular backbone that it is associated with is set
with the \code{y} aesthetic.
}
\section{Aesthetics}{

\itemize{
\item xmin,xmax (required; start and end positions of the helix)
\item y (required; the molecular backbone)
\item alpha
\item colour
\item linetype
\item linewidth
}
}

\examples{

ggplot2::ggplot(subset(demolin_ss, type == "helix"), 
                aes(xmin = start, xmax = end, y = 1)) +
  geom_helix(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_helix_label]{geom_helix_label()}}
}
